Updated Pages to new ViewModel
