+++ ContentPage/ClientPage.xaml.cs
@@ -1,44 +1,24 @@
-﻿using System.Windows.Controls;
-using Analyzer;
-using Content.Server;
+﻿using Content.ViewModel;
 using Networking.Communicator;
-
-using System;
-using System.Collections.Generic;
-using System.Windows;
+using System.Windows.Controls;
 
 
 namespace ContentPage
 {
+    /// <summary>
+    /// Interaction logic for ClientPage.xaml
+    /// </summary>
     public partial class ClientPage : Page
     {
-        private ContentServerViewModel viewModel;
-
-        public ClientPage(ICommunicator server)
+        private ContentClientViewModel viewModel;
+        public ClientPage(ICommunicator client, string sessionID)
         {
             InitializeComponent();
-            viewModel = new ContentServerViewModel(server);
+            viewModel = new ContentClientViewModel(client, sessionID);
             DataContext = viewModel;
 
-            LoadResultPage(); // Load ResultPage initially
-            LoadConfigurationPage(); // Optionally, load ConfigurationPage initially
-        }
-
-       
-
-        private void LoadResultPage()
-        {
-            ResultPage resultPage = new ResultPage(viewModel);
-            ResultFrame.NavigationService.Navigate(resultPage);
-            
-        }
-
-        private void LoadConfigurationPage()
-        {
-            ConfigurationPage configPage = new ConfigurationPage(viewModel);
-            ConfigFrame.NavigationService.Navigate(configPage);
-            
+            UploadFrame.Navigate(new FileUpload(viewModel));
+            ResultFrame.Navigate(new ResultPage(viewModel));
         }
     }
 }
-
