+++ ContentPage/ServerPage.xaml
@@ -1,4 +1,4 @@
-﻿<Page x:Class="ContentPage.ClientPage"
+﻿<Page x:Class="ContentPage.ServerPage"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -11,10 +11,11 @@
     <Grid>
 
         <StackPanel Orientation="Vertical">
-            <Frame x:Name="UploadFrame" />
+            <Frame x:Name="ConfigFrame" />
             <Frame x:Name="ResultFrame" />
-
+            
         </StackPanel>
+        
 
     </Grid>
 </Page>
